Intermediate Progress on Qt6 Migration

+++ include/xstudio/ui/qml/helper_ui.hpp
@@ -442,7 +442,7 @@ namespace ui {
 
             Q_INVOKABLE [[nodiscard]] QColor
             saturate(const QColor &color, const double factor = 1.5) const {
-                double h, s, l, a;
+                float h, s, l, a;
                 color.getHslF(&h, &s, &l, &a);
                 s = std::max(0.0, std::min(1.0, s * factor));
                 return QColor::fromHslF(h, s, l, a);
@@ -450,7 +450,7 @@ namespace ui {
 
             Q_INVOKABLE [[nodiscard]] QColor
             luminate(const QColor &color, const double factor = 1.5) const {
-                double h, s, l, a;
+                float h, s, l, a;
                 color.getHslF(&h, &s, &l, &a);
                 l = std::max(0.0, std::min(1.0, l * factor));
                 return QColor::fromHslF(h, s, l, a);
@@ -467,7 +467,7 @@ namespace ui {
                 const QColor &color,
                 const double sfactor = 1.0,
                 const double lfactor = 1.0) const {
-                double h, s, l, a;
+                float h, s, l, a;
                 color.getHslF(&h, &s, &l, &a);
                 s = std::max(0.0, std::min(1.0, s * sfactor));
                 l = std::max(0.0, std::min(1.0, l * lfactor));
